Crate pallet_interest_accrual

Interest Accrual Pallet

A pallet for calculating interest accrual on debt. It keeps track of different buckets of interest rates and is optimized for many loans per interest bucket. This implementation is inspired by jug.sol from Multi Collateral Dai.

It works by defining debt = normalized debt * accumulated rate. When the first loan for an interest rate is created, the accumulated rate is set to 1.0. The normalized debt is then calculated, which is the debt at t=0, using debt / accumulated rate.

Over time, the accumulated rate grows based on the interest rate per second. Any time the accumulated rate is updated for an interest rate group, this indirectly updates the debt of all loans outstanding using this interest rate.

                   ar = accumulated rate
                   nd = normalized debt
       
           │
       2.0 │                             ****
           │                         ****
           │                     ****
           │                 ****
  ar   1.5 │             ****
           │         ****
           │      ****
           │   ****
       1.0 │ **
           └──────────────────────────────────
           │              │
                            
           borrow 10      borrow 20
           ar   = 1.0     ar   = 1.5
           nd   = 10      nd   = 10 + (20 / 1.5) = 23.33
           debt = 10      debt = 35

Re-exports

pub use pallet::*;

Modules

migration

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

weights

Autogenerated weights for pallet_interest_accrual

Structs

RateDetails

RateDetailsV0

Enums

Release