Type Alias pallet_interest_accrual::pallet::Module

pub type Module<T> = Pallet<T>;

๐Deprecated: use Pallet instead

Type alias to Pallet, to be used by construct_runtime.

Generated by pallet attribute macro.

Aliased Type

struct Module<T>(/* private fields */);

Implementations

impl<T: Config> Pallet<T>

pub fn get_debt( interest_rate_per_year: &InterestRate<T::Rate>, normalized_debt: T::Balance, when: Seconds ) -> Result<T::Balance, DispatchError>

Calculate fastly the current debt using normalized debt * cumulative rate if when is exactly now (same block). If when is in the past it recomputes the previous cumulative rate.

If when is further in the past than the last time the normalized debt was adjusted, this will return nonsense (effectively โrewinding the clockโ to before the value was valid)

pub fn do_adjust_normalized_debt( interest_rate_per_year: &InterestRate<T::Rate>, normalized_debt: T::Balance, adjustment: Adjustment<T::Balance> ) -> Result<T::Balance, DispatchError>

pub fn do_renormalize_debt( old_interest_rate: &InterestRate<T::Rate>, new_interest_rate: &InterestRate<T::Rate>, normalized_debt: T::Balance ) -> Result<T::Balance, DispatchError>

pub fn calculate_accumulated_rate<Rate: FixedPointNumber>( interest_rate_per_sec: Rate, accumulated_rate: Rate, last_updated: Seconds, now: Seconds ) -> Result<Rate, ArithmeticError>

pub fn reference_interest_rate( interest_rate_per_year: &InterestRate<T::Rate> ) -> DispatchResult

pub fn unreference_interest_rate( interest_rate_per_year: &InterestRate<T::Rate> ) -> DispatchResult

pub fn get_rate( interest_rate_per_year: &InterestRate<T::Rate> ) -> Result<RateDetails<<T as Config>::Rate>, DispatchError>

impl<T: Config> Pallet<T>

pub fn rates() -> BoundedVec<RateDetails<<T as Config>::Rate>, T::MaxRateCount>

impl<T: Config> Pallet<T>

pub fn last_updated() -> Seconds